Implement real stat collection and updating
The stats for average records added per hour, and the ratio of average records
added versus deleted per hour are collected and displayed.

Note that a mock/in-browser backend is still being used; as opposed to
hot reloading just the statistics upon an addition or deletion of a
record, the mock DB is polled every second by the client to check for
updates.